fix(appearance): apply the makeover selection exactly as picked
The makeover confirm was clamped against stale identikit and colour
ranges, so modern cache kits snapped back to defaults and leg/torso
colour indices above the old bounds reset to 0.

- send kit ids as signed shorts and length APPEARANCE_SET for it, since
  current caches use ids past 255 and 255 itself is a valid kit
- drop the fixed kit/colour clamps; the cache owns those ranges
- keep the torso's cache face priorities so pockets/overlays stack, and
  spread them across layers so the front overlay gets real depth margin
- start each PlayerDesign (679) open from the live appearance instead of
  a stale editing session

Fixes #88

+++ client/network/serverConnection/outgoing/widgets.ts
@@ -203,15 +203,17 @@ export function sendPlayerDesignConfirm(appearance: {
     const { createPacket, queuePacket } = require("../../packet");
     const { ClientPacketId } = require("../../../common/network/ClientPacketId");
 
-    // Payload: gender (1), kits[7] (7, -1=0xff), colors[5] (5)
+    // Payload: gender (1), kits[7] (14, signed short, -1=0xffff), colors[5] (5)
     const pkt = createPacket(ClientPacketId.APPEARANCE_SET);
     const gender = (appearance.gender | 0) === 1 ? 1 : 0;
     pkt.packetBuffer.writeByte(gender);
 
+    // Identity kit ids in current caches exceed 255, and 255 is itself a valid kit, so kits
+    // must go out as shorts rather than bytes (a byte would alias 256-306 and steal the -1 sentinel).
     const kits = Array.isArray(appearance.kits) ? appearance.kits : [];
     for (let i = 0; i < 7; i++) {
         const v = Number.isFinite(kits[i]) ? kits[i] | 0 : -1;
-        pkt.packetBuffer.writeByte(v);
+        pkt.packetBuffer.writeShort(v);
     }
 
     const colors = Array.isArray(appearance.colors) ? appearance.colors : [];
